Write a C program that prints the size of various types on the computer it is compiled and run on

+++ 0x00-hello_world/6-size.c
@@ -8,7 +8,7 @@ int main(void)
 {
   char charType;
   int intType;
-  long longType;
+  long long int  longType;
   float floatType;
 
   /*sizeof evaluates size of variable*/
